third_party/blink/web_tests/fast/dom/shadow/input-shadow-nochange-expected.txt

This test checks input's shadow root is not changed if input type is changed.

hidden
PASS newShadow is shadow
text
PASS newShadow is shadow
search
PASS newShadow is shadow
tel
PASS newShadow is shadow
url
PASS newShadow is shadow
email
PASS newShadow is shadow
password
PASS newShadow is shadow
datetime
PASS newShadow is shadow
date
PASS newShadow is shadow
month
PASS newShadow is shadow
week
PASS newShadow is shadow
time
PASS newShadow is shadow
datetime-local
PASS newShadow is shadow
number
PASS newShadow is shadow
range
PASS newShadow is shadow
color
PASS newShadow is shadow
checkbox
PASS newShadow is shadow
radio
PASS newShadow is shadow
file
PASS newShadow is shadow
submit
PASS newShadow is shadow
image
PASS newShadow is shadow
reset
PASS newShadow is shadow
button
PASS newShadow is shadow
PASS successfullyParsed is true

TEST COMPLETE
